+++ src/test/java/kitchenpos/products/tobe/domain/ProductTest.java
@@ -3,6 +3,8 @@ package kitchenpos.products.tobe.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 class ProductTest {
@@ -12,7 +14,7 @@ class ProductTest {
     void create() {
         // given
         final String name = "새로운 제품";
-        final Long price = 1000L;
+        final BigDecimal price = BigDecimal.valueOf(1000);
 
         // when
         Product product = new Product(name, price);
